split quantize_row_q8_0/1
